chat_ws/requirements.txt
aioredis==1.3.1
amqp==5.0.5
asgiref==3.3.1
async-timeout==3.0.1
attrs==20.3.0
autobahn==21.3.1
Automat==20.2.0
billiard==3.6.3.0
celery==5.0.5
certifi==2020.12.5
cffi==1.14.5
channels==3.0.3
channels-redis==3.2.0
chardet==4.0.0
click==7.1.2
click-didyoumean==0.0.3
click-plugins==1.1.1
click-repl==0.1.6
constantly==15.1.0
cryptography==3.4.7
daphne==3.0.1
Django==3.1.7
django-channels-jwt-auth-middleware==1.0.0
django-cors-headers==3.7.0
djangorestframework==3.12.4
djangorestframework-simplejwt==4.6.0
hiredis==2.0.0
hyperlink==21.0.0
idna==2.10
incremental==21.3.0
kombu==5.0.2
msgpack==1.0.2
prompt-toolkit==3.0.18
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
PyJWT==2.0.1
pyOpenSSL==20.0.1
pytz==2021.1
redis==3.5.3
requests==2.25.1
service-identity==18.1.0
six==1.15.0
sqlparse==0.4.1
Twisted==21.2.0
txaio==21.2.1
urllib3==1.26.4
vine==5.0.0
wcwidth==0.2.5
zope.interface==5.3.0
